% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_describe_hsm}
\alias{cloudhsm_describe_hsm}
\title{This is documentation for AWS CLOUDHSM CLASSIC}
\usage{
cloudhsm_describe_hsm(HsmArn, HsmSerialNumber)
}
\arguments{
\item{HsmArn}{The ARN of the HSM. Either the \code{HsmArn} or the \code{SerialNumber} parameter
must be specified.}

\item{HsmSerialNumber}{The serial number of the HSM. Either the \code{HsmArn} or the
\code{HsmSerialNumber} parameter must be specified.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Retrieves information about an HSM. You can identify the HSM by its ARN
or its serial number.
}
\section{Request syntax}{
\preformatted{svc$describe_hsm(
  HsmArn = "string",
  HsmSerialNumber = "string"
)
}
}

\keyword{internal}
