% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool_domain}
\alias{cognitoidentityprovider_delete_user_pool_domain}
\title{Deletes a domain for a user pool}
\usage{
cognitoidentityprovider_delete_user_pool_domain(Domain, UserPoolId)
}
\arguments{
\item{Domain}{[required] The domain string.}

\item{UserPoolId}{[required] The user pool ID.}
}
\description{
Deletes a domain for a user pool.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$delete_user_pool_domain(
  Domain = "string",
  UserPoolId = "string"
)
}
}

\keyword{internal}
