% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_pool_domain}
\alias{cognitoidentityprovider_describe_user_pool_domain}
\title{Gets information about a domain}
\usage{
cognitoidentityprovider_describe_user_pool_domain(Domain)
}
\arguments{
\item{Domain}{[required] The domain string.}
}
\description{
Gets information about a domain.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$describe_user_pool_domain(
  Domain = "string"
)
}
}

\keyword{internal}
