% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_devices}
\alias{cognitoidentityprovider_list_devices}
\title{Lists the devices}
\usage{
cognitoidentityprovider_list_devices(AccessToken, Limit,
  PaginationToken)
}
\arguments{
\item{AccessToken}{[required] The access tokens for the request to list devices.}

\item{Limit}{The limit of the device request.}

\item{PaginationToken}{The pagination token for the list request.}
}
\description{
Lists the devices.
}
\section{Request syntax}{
\preformatted{svc$list_devices(
  AccessToken = "string",
  Limit = 123,
  PaginationToken = "string"
)
}
}

\keyword{internal}
