% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_directories}
\alias{directoryservice_describe_directories}
\title{Obtains information about the directories that belong to this account}
\usage{
directoryservice_describe_directories(DirectoryIds, NextToken, Limit)
}
\arguments{
\item{DirectoryIds}{A list of identifiers of the directories for which to obtain the
information. If this member is null, all directories that belong to the
current account are returned.

An empty list results in an \code{InvalidParameterException} being thrown.}

\item{NextToken}{The \code{DescribeDirectoriesResult.NextToken} value from a previous call to
DescribeDirectories. Pass null if this is the first call.}

\item{Limit}{The maximum number of items to return. If this value is zero, the
maximum number of items is specified by the limitations of the
operation.}
}
\description{
Obtains information about the directories that belong to this account.
}
\details{
You can retrieve information about specific directories by passing the
directory identifiers in the \code{DirectoryIds} parameter. Otherwise, all
directories that belong to the current account are returned.

This operation supports pagination with the use of the \code{NextToken}
request and response parameters. If more results are available, the
\code{DescribeDirectoriesResult.NextToken} member contains a token that you
pass in the next call to DescribeDirectories to retrieve the next set of
items.

You can also specify a maximum number of return results with the \code{Limit}
parameter.
}
\section{Request syntax}{
\preformatted{svc$describe_directories(
  DirectoryIds = list(
    "string"
  ),
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
