% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_get_directory_limits}
\alias{directoryservice_get_directory_limits}
\title{Obtains directory limit information for the current region}
\usage{
directoryservice_get_directory_limits()
}
\description{
Obtains directory limit information for the current region.
}
\section{Request syntax}{
\preformatted{svc$get_directory_limits()
}
}

\keyword{internal}
