% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_decline_invitations}
\alias{guardduty_decline_invitations}
\title{Declines invitations sent to the current member account by AWS account
specified by their account IDs}
\usage{
guardduty_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to decline invitations from.}
}
\description{
Declines invitations sent to the current member account by AWS account specified by their account IDs.
}
\section{Request syntax}{
\preformatted{guardduty$decline_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
