% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_detector}
\alias{guardduty_delete_detector}
\title{Deletes a Amazon GuardDuty detector specified by the detector ID}
\usage{
guardduty_delete_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID that specifies the detector that you want to delete.}
}
\description{
Deletes a Amazon GuardDuty detector specified by the detector ID.
}
\section{Request syntax}{
\preformatted{svc$delete_detector(
  DetectorId = "string"
)
}
}

\keyword{internal}
