% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_filter}
\alias{guardduty_delete_filter}
\title{Deletes the filter specified by the filter name}
\usage{
guardduty_delete_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{[required] The unique ID that specifies the detector where you want to delete a filter.}

\item{FilterName}{[required] The name of the filter.}
}
\description{
Deletes the filter specified by the filter name.
}
\section{Request syntax}{
\preformatted{guardduty$delete_filter(
  DetectorId = "string",
  FilterName = "string"
)
}
}

\keyword{internal}
