% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_threat_intel_set}
\alias{guardduty_get_threat_intel_set}
\title{Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID}
\usage{
guardduty_get_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose ThreatIntelSet you want to describe.}

\item{ThreatIntelSetId}{[required] The unique ID that specifies the ThreatIntelSet that you want to describe.}
}
\description{
Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID.
}
\section{Request syntax}{
\preformatted{guardduty$get_threat_intel_set(
  DetectorId = "string",
  ThreatIntelSetId = "string"
)
}
}

\keyword{internal}
