% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_invitations}
\alias{guardduty_list_invitations}
\title{Lists all GuardDuty membership invitations that were sent to the current
AWS account}
\usage{
guardduty_list_invitations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{You can use this parameter to indicate the maximum number of invitations you want in the response. The default value is 50. The maximum value is 50.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListInvitations action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.}
}
\description{
Lists all GuardDuty membership invitations that were sent to the current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
