% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_threat_intel_sets}
\alias{guardduty_list_threat_intel_sets}
\title{Lists the ThreatIntelSets of the GuardDuty service specified by the
detector ID}
\usage{
guardduty_list_threat_intel_sets(DetectorId, MaxResults, NextToken)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose ThreatIntelSets you want to list.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 7. The maximum value is 7.}

\item{NextToken}{Pagination token to start retrieving threat intel sets from.}
}
\description{
Lists the ThreatIntelSets of the GuardDuty service specified by the detector ID.
}
\section{Request syntax}{
\preformatted{svc$list_threat_intel_sets(
  DetectorId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
