% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_get_resource_policies}
\alias{ram_get_resource_policies}
\title{Gets the policies for the specifies resources}
\usage{
ram_get_resource_policies(resourceArns, principal, nextToken,
  maxResults)
}
\arguments{
\item{resourceArns}{[required] The Amazon Resource Names (ARN) of the resources.}

\item{principal}{The principal.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Gets the policies for the specifies resources.
}
\section{Request syntax}{
\preformatted{svc$get_resource_policies(
  resourceArns = list(
    "string"
  ),
  principal = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
