% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_disable_standards}
\alias{securityhub_batch_disable_standards}
\title{Disables the standards specified by the standards subscription ARNs}
\usage{
securityhub_batch_disable_standards(StandardsSubscriptionArns)
}
\arguments{
\item{StandardsSubscriptionArns}{[required] The ARNS of the standards subscriptions that you want to disable.}
}
\description{
Disables the standards specified by the standards subscription ARNs. In
the context of Security Hub, supported standards (for example, CIS AWS
Foundations) are automated and continuous checks that help determine
your compliance status against security industry (including AWS) best
practices.
}
\section{Request syntax}{
\preformatted{securityhub$batch_disable_standards(
  StandardsSubscriptionArns = list(
    "string"
  )
)
}
}

\keyword{internal}
