% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_members}
\alias{securityhub_delete_members}
\title{Deletes the Security Hub member accounts that are specified by the
account IDs}
\usage{
securityhub_delete_members(AccountIds)
}
\arguments{
\item{AccountIds}{A list of account IDs of the Security Hub member accounts that you want
to delete.}
}
\description{
Deletes the Security Hub member accounts that are specified by the
account IDs.
}
\section{Request syntax}{
\preformatted{svc$delete_members(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
