% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_security_hub}
\alias{securityhub_disable_security_hub}
\title{Disables the AWS Security Hub Service}
\usage{
securityhub_disable_security_hub()
}
\description{
Disables the AWS Security Hub Service.
}
\section{Request syntax}{
\preformatted{svc$disable_security_hub()
}
}

\keyword{internal}
