% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_members}
\alias{securityhub_disassociate_members}
\title{Disassociates the Security Hub member accounts that are specified by the
account IDs from their master account}
\usage{
securityhub_disassociate_members(AccountIds)
}
\arguments{
\item{AccountIds}{The account IDs of the member accounts that you want to disassociate
from the master account.}
}
\description{
Disassociates the Security Hub member accounts that are specified by the
account IDs from their master account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_members(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
