% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_security_hub}
\alias{securityhub_enable_security_hub}
\title{Enables the AWS Security Hub service}
\usage{
securityhub_enable_security_hub()
}
\description{
Enables the AWS Security Hub service.
}
\section{Request syntax}{
\preformatted{svc$enable_security_hub()
}
}

\keyword{internal}
