% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_insights}
\alias{securityhub_get_insights}
\title{Lists and describes insights that are specified by insight ARNs}
\usage{
securityhub_get_insights(InsightArns, NextToken, MaxResults)
}
\arguments{
\item{InsightArns}{The ARNS of the insights that you want to describe.}

\item{NextToken}{Paginates results. Set the value of this parameter to NULL on your first
call to the GetInsights operation. For subsequent calls to the
operation, fill nextToken in the request with the value of nextToken
from the previous response to continue listing data.}

\item{MaxResults}{Indicates the maximum number of items that you want in the response.}
}
\description{
Lists and describes insights that are specified by insight ARNs.
}
\section{Request syntax}{
\preformatted{svc$get_insights(
  InsightArns = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
