% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_create_subscription}
\alias{shield_create_subscription}
\title{Activates AWS Shield Advanced for an account}
\usage{
shield_create_subscription()
}
\description{
Activates AWS Shield Advanced for an account.
}
\details{
As part of this request you can specify \code{EmergencySettings} that
automaticaly grant the DDoS response team (DRT) needed permissions to
assist you during a suspected DDoS attack. For more information see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/authorize-DRT.html}{Authorize the DDoS Response Team to Create Rules and Web ACLs on Your Behalf}.

When you initally create a subscription, your subscription is set to be
automatically renewed at the end of the existing subscription period.
You can change this by submitting an \code{UpdateSubscription} request.
}
\section{Request syntax}{
\preformatted{shield$create_subscription()
}
}

\keyword{internal}
