% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_list_protections}
\alias{shield_list_protections}
\title{Lists all Protection objects for the account}
\usage{
shield_list_protections(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The \code{ListProtectionsRequest.NextToken} value from a previous call to
\code{ListProtections}. Pass null if this is the first call.}

\item{MaxResults}{The maximum number of Protection objects to be returned. If this is left
blank the first 20 results will be returned.

This is a maximum value; it is possible that AWS WAF will return the
results in smaller batches. That is, the number of Protection objects
returned could be less than \code{MaxResults}, even if there are still more
Protection objects yet to return. If there are more Protection objects
to return, AWS WAF will always also return a \code{NextToken}.}
}
\description{
Lists all Protection objects for the account.
}
\section{Request syntax}{
\preformatted{shield$list_protections(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
