% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_rate_based_rule}
\alias{waf_get_rate_based_rule}
\title{Returns the RateBasedRule that is specified by the RuleId that you
included in the GetRateBasedRule request}
\usage{
waf_get_rate_based_rule(RuleId)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the RateBasedRule that you want to get. \code{RuleId} is
returned by CreateRateBasedRule and by ListRateBasedRules.}
}
\description{
Returns the RateBasedRule that is specified by the \code{RuleId} that you
included in the \code{GetRateBasedRule} request.
}
\section{Request syntax}{
\preformatted{svc$get_rate_based_rule(
  RuleId = "string"
)
}
}

\keyword{internal}
