% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_list_subscribed_rule_groups}
\alias{waf_list_subscribed_rule_groups}
\title{Returns an array of RuleGroup objects that you are subscribed to}
\usage{
waf_list_subscribed_rule_groups(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more
\code{ByteMatchSets}subscribed rule groups than the value of \code{Limit}, AWS WAF
returns a \code{NextMarker} value in the response that allows you to list
another group of subscribed rule groups. For the second and subsequent
\code{ListSubscribedRuleGroupsRequest} requests, specify the value of
\code{NextMarker} from the previous response to get information about another
batch of subscribed rule groups.}

\item{Limit}{Specifies the number of subscribed rule groups that you want AWS WAF to
return for this request. If you have more objects than the number you
specify for \code{Limit}, the response includes a \code{NextMarker} value that you
can use to get another batch of objects.}
}
\description{
Returns an array of RuleGroup objects that you are subscribed to.
}
\section{Request syntax}{
\preformatted{svc$list_subscribed_rule_groups(
  NextMarker = "string",
  Limit = 123
)
}
}

\keyword{internal}
